+++ src/components/Login/index.js
@@ -1,0 +1,2 @@
+import Login from './Login';
+export default Login;
