+++ src/components/Products/index.js
@@ -1,0 +1,2 @@
+import Product from './Products';
+export default Product;
