+++ src/components/User/User.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { exchangeToken, logout } from '../state/actionCreators/authAC';
+import { exchangeToken, logout } from '../../state/actionCreators/authAC';
 import { Navigate } from 'react-router-dom';
 
 function User() {
