+++ src/components/User/index.js
@@ -1,0 +1,2 @@
+import User from './User'
+export default User
